pyproject.toml
[tool.poetry]
name = "diary-project"
version = "0.1.0"
description = "나만의 일기장 프로젝트"
authors = ["빠르게가조 <Fastgo@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
asyncpg = "^0.30.0"
pydantic = "^2.9.0"
python-dotenv = "^1.0.1"
passlib = { version = "^1.7.4", extras = ["bcrypt"] }
python-multipart = "^0.0.9"
aiohttp = "^3.8.1"
beautifulsoup4 = "^4.12.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
httpx = "^0.28.0"
ruff = "^0.6.4"
black = "^24.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

